+++ contracts/l2/v1/impl/Storage.sol
@@ -2,7 +2,7 @@
 
 pragma solidity ^0.8.4;
 
-import {Types} from "../lib/Types.sol";
+import { Types } from "../lib/Types.sol";
 
 contract Storage {
     // token id => token address(erc20), reduce storage
@@ -15,8 +15,10 @@ contract Storage {
     mapping(address => address) internal _bindingAccount;
 
     // position id => position
-    mapping(uint256 => Types.Position) internal _position;
+    mapping(uint64 => Types.Position) internal _position;
 
     // order id => order
     mapping(uint256 => Types.Order) internal _order;
+
+    // TODO:不同的合约不同的配置参数，价格
 }
